refactor(module-infra): 重构数据访问层
- 新增 ApiAccessLogDataRepository、ApiErrorLogDataRepository、ConfigDataRepository、FileContentDataRepository 和 FileDataRepository- 更新相关服务类，使用新的数据仓库类替换原有的 DataRepository- 优化日志清理和分页查询等操作的实现
- 调整 FileClient 的实现，使用新的 FileContentDataRepository 方法

+++ onebase-framework/onebase-spring-boot-starter-orm/src/main/java/com/cmsr/onebase/framework/aynline/DataRepositoryNew.java
@@ -44,8 +44,7 @@ public class DataRepositoryNew<T extends BaseDO> {
 
     private Class<T> defaultClazz = null;
 
-    // public DataRepositoryNew() {
-    // }
+    public static boolean isDebug = false;
 
     public DataRepositoryNew(Class<T> defaultClazz) {
         this.defaultClazz = defaultClazz;
@@ -155,7 +154,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             ConfigStore configs = new DefaultConfigStore();
             configs.eq(BaseDO.ID, entity.getId());
             Long result = anylineService.update(entity, configs);
-            log.info("update  ---> class={}, effect rows = {}", entity.getClass().getSimpleName(), result);
+            if (isDebug) {
+                log.info("update  ---> class={}, effect rows = {}", entity.getClass().getSimpleName(), result);
+            }
             return result;
         } catch (Exception e) {
             log.error("update error, class={}, entity={}", defaultClazz, entity, e);
@@ -195,7 +196,9 @@ public class DataRepositoryNew<T extends BaseDO> {
         }
         try {
             long result = anylineService.update(1000, getTableName(entity.getClass()), entity, configs, (List<String>) null);
-            log.info("updateByConfig  ---> class={}, effect rows = {}", entity.getClass().getSimpleName(), result);
+            if (isDebug) {
+                log.info("updateByConfig  ---> class={}, effect rows = {}", entity.getClass().getSimpleName(), result);
+                }
             return result;
         } catch (Exception e) {
             log.error("updateByConfig error, class={}, entity={}, config={}", defaultClazz, entity, configs, e);
@@ -324,7 +327,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             ConfigStore configs = new DefaultConfigStore();
             String tableName = getTableName(defaultClazz);
             DataSet dataSet = anylineService.querys(tableName, configs);
-            log.error("findAll --->  dataSet = {}", dataSet);
+            if (isDebug) {
+                log.info("findAll --->  dataSet.size = {}", dataSet.size());
+            }
             return dataSet.entity(defaultClazz);
         } catch (Exception e) {
             log.error("findAll error, class={}", defaultClazz, e);
@@ -362,9 +367,6 @@ public class DataRepositoryNew<T extends BaseDO> {
             return Collections.emptyList();
         }
         try {
-            if (ids == null || ids.isEmpty()) {
-                return Collections.emptyList();
-            }
 
             ConfigStore configs = new DefaultConfigStore();
             configs.in(BaseDO.ID, ids);
@@ -450,7 +452,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             DataRow row = new DataRow();
             row.put(BaseDO.DELETED, System.currentTimeMillis());  // 设置逻辑删除标记
             long result = anylineService.update(getTableName(defaultClazz), row, configs);
-            log.info("deleteByConfig  ---> class={}, effect rows = {}", defaultClazz, result);
+            if (isDebug) {
+                log.info("deleteByConfig  ---> class={}, effect rows = {}", defaultClazz, result);
+            }
             return result;
         } catch (Exception e) {
             log.error("deleteByConfig error, ---> class={}, configs={}", defaultClazz.getSimpleName(), configs, e);
@@ -475,7 +479,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             DataRow row = new DataRow();
             row.put(BaseDO.DELETED, System.currentTimeMillis());  // 设置逻辑删除标记
             long result = anylineService.update(getTableName(defaultClazz), row, configs);
-            log.info("deleteById  ---> class={}, effect rows = {}, id = {}", defaultClazz, result, id);
+            if (isDebug) {
+                log.info("deleteById  ---> class={}, effect rows = {}, id = {}", defaultClazz, result, id);
+            }
             return result;
         } catch (Exception e) {
             log.error("deleteById error, class={}, id={}", defaultClazz.getSimpleName(), id, e);
@@ -525,7 +531,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             DataRow row = new DataRow();
             row.put(BaseDO.DELETED, System.currentTimeMillis());  // 设置逻辑删除标记
             long result = anylineService.update(getTableName(defaultClazz), row, configs);
-            log.info("deleteAllById  ---> class={}, effect rows={}, ids={}", defaultClazz, result, ids);
+            if (isDebug) {
+                log.info("deleteAllById  ---> class={}, effect rows={}, ids={}", defaultClazz, result, ids);
+            }
             return result;
         } catch (Exception e) {
             log.error("deleteAllById error, class={}, ids={}", defaultClazz.getSimpleName(), ids, e);
@@ -544,8 +552,9 @@ public class DataRepositoryNew<T extends BaseDO> {
             DataRow row = new DataRow();
             row.put(BaseDO.DELETED, System.currentTimeMillis());  // 设置逻辑删除标记
             long result = anylineService.update(getTableName(defaultClazz), row, configs);
-            log.info("deleteAll  ---> class={}, effect rows={}", defaultClazz, result);
-
+            if (isDebug) {
+                log.info("deleteAll  ---> class={}, effect rows={}", defaultClazz, result);
+            }
         } catch (Exception e) {
             log.error("deleteAll error, class={}", defaultClazz.getSimpleName(), e);
             throw new BizException(StatusCode.DB_DELETE_ERROR);
